refactor: models dir.
