+++ src/index.ts
@@ -8,6 +8,14 @@ export function convertToDbioUnitString(_number: number): string {
   return (_number * dbioUnit).toString();
 }
 
+export function convertSubstrateNumberToNumber(data: any): number {
+  return Number(data.toString().split(',').join(''));
+}
+
+export function convertSubstrateBalanceToNumber(bal: any): number {
+  return convertSubstrateNumberToNumber(bal) / dbioUnit;
+}
+
 export * from './query';
 export * from './models';
 export * from './command';
