+++ src/models/index.ts
@@ -1,6 +1,3 @@
 export * from './labs';
-export * from './services';
-export * from './certifications';
-export * from './orders';
-export * from './genetic-testing';
+export * from './genetic-analysts';
 export * from './service-request';
