+++ src/models/labs/genetic-testing/index.ts
